Keep cancel field on disk #212 (#215)

+++ jobqueue/src/main/java/com/birbit/android/jobqueue/persistentQueue/sqlite/DbOpenHelper.java
@@ -8,7 +8,7 @@ import android.database.sqlite.SQLiteOpenHelper;
  * Helper class for {@link SqliteJobQueue} to handle database connection
  */
 public class DbOpenHelper extends SQLiteOpenHelper {
-    private static final int DB_VERSION = 11;
+    private static final int DB_VERSION = 12;
     /*package*/ static final String JOB_HOLDER_TABLE_NAME = "job_holder";
     /*package*/ static final String JOB_TAGS_TABLE_NAME = "job_holder_tags";
     /*package*/ static final SqlHelper.Property INSERTION_ORDER_COLUMN = new SqlHelper.Property("insertionOrder", "integer", 0);
@@ -22,6 +22,7 @@ public class DbOpenHelper extends SQLiteOpenHelper {
     /*package*/ static final SqlHelper.Property REQUIRED_NETWORK_TYPE_OLUMN = new SqlHelper.Property("network_type", "integer", 8);
     /*package*/ static final SqlHelper.Property DEADLINE_COLUMN = new SqlHelper.Property("deadline", "integer", 9);
     /*package*/ static final SqlHelper.Property CANCEL_ON_DEADLINE_COLUMN = new SqlHelper.Property("cancel_on_deadline", "integer", 10);
+    /*package*/ static final SqlHelper.Property CANCELLED_COLUMN = new SqlHelper.Property("cancelled", "integer", 11);
 
     /*package*/ static final SqlHelper.Property TAGS_ID_COLUMN = new SqlHelper.Property("_id", "integer", 0);
     /*package*/ static final SqlHelper.Property TAGS_JOB_ID_COLUMN = new SqlHelper.Property("job_id", "text", 1, new SqlHelper.ForeignKey(JOB_HOLDER_TABLE_NAME, ID_COLUMN.columnName));
@@ -29,7 +30,7 @@ public class DbOpenHelper extends SQLiteOpenHelper {
 
 
 
-    /*package*/ static final int COLUMN_COUNT = 11;
+    /*package*/ static final int COLUMN_COUNT = 12;
     /*package*/ static final int TAGS_COLUMN_COUNT = 3;
 
     static final String TAG_INDEX_NAME = "TAG_NAME_INDEX";
@@ -51,10 +52,10 @@ public class DbOpenHelper extends SQLiteOpenHelper {
                 RUNNING_SESSION_ID_COLUMN,
                 REQUIRED_NETWORK_TYPE_OLUMN,
                 DEADLINE_COLUMN,
-                CANCEL_ON_DEADLINE_COLUMN
+                CANCEL_ON_DEADLINE_COLUMN,
+                CANCELLED_COLUMN
         );
         sqLiteDatabase.execSQL(createQuery);
-
         String createTagsQuery = SqlHelper.create(JOB_TAGS_TABLE_NAME,
                 TAGS_ID_COLUMN,
                 TAGS_JOB_ID_COLUMN,
@@ -67,10 +68,20 @@ public class DbOpenHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion) {
-        sqLiteDatabase.execSQL(SqlHelper.drop(JOB_HOLDER_TABLE_NAME));
-        sqLiteDatabase.execSQL(SqlHelper.drop(JOB_TAGS_TABLE_NAME));
-        sqLiteDatabase.execSQL("DROP INDEX IF EXISTS " + TAG_INDEX_NAME);
-        onCreate(sqLiteDatabase);
+        if (oldVersion == 11) {
+            addCancelColumn(sqLiteDatabase);
+        } else {
+            sqLiteDatabase.execSQL(SqlHelper.drop(JOB_HOLDER_TABLE_NAME));
+            sqLiteDatabase.execSQL(SqlHelper.drop(JOB_TAGS_TABLE_NAME));
+            sqLiteDatabase.execSQL("DROP INDEX IF EXISTS " + TAG_INDEX_NAME);
+            onCreate(sqLiteDatabase);
+        }
+    }
+
+    private void addCancelColumn(SQLiteDatabase sqLiteDatabase) {
+        String query = "ALTER TABLE " + JOB_HOLDER_TABLE_NAME + " ADD COLUMN "
+                + CANCELLED_COLUMN.columnName + " " + CANCELLED_COLUMN.type;
+        sqLiteDatabase.execSQL(query);
     }
 
     @Override
